fix: Replace text in brackets
This improves on previous attempt where tried to identify team by the 'team' column which did not work because it would say things like 'Monaco' whereas in the commentary, they would use 'AS Monaco'.

This assumes that teams players belong to are always in brackets and more crucially, that there is no other important text in brackets.